+++ doc/release_notes.rst
@@ -72,6 +72,8 @@ PyPSA-Earth 0.8.0
 
 **Minor Changes and bug-fixing**
 
+* Fix issues where gas and biomass powerplant capacity were not added in ``add_existing_baseyear.py``. Differentiate `urban central solid biomass CHP` with non-CHP `biomass` for historical capacity `PR #1678 <https://github.com/pypsa-meets-earth/pypsa-earth/pull/1678>`__
+
 * Avoid creating duplicate conventional generators by setting `keep_existing_capacities` to `false`. Convert CCGT as links in `prepare_sector_network` `PR #1630 <https://github.com/pypsa-meets-earth/pypsa-earth/pull/1630>`__
 
 * Refine load shedding capacity calculation to use bus-specific maximum loads instead of fixed large values, improving solver performance and numeric stability. Rename load shedding generator carrier to 'load shedding' (now) instead of just 'load' (previously) `PR #1581 <https://github.com/pypsa-meets-earth/pypsa-earth/pull/1581>`__